+++ pyxel/calibration/__init__.py
@@ -19,6 +19,8 @@ from .util import (
     check_ranges,
     list_to_slice,
     read_data,
+    read_datacubes,
+    list_to_3d_slice,
 )
 from .archipelago import MyArchipelago
 from .calibration import Calibration, CalibrationMode