+++ pyxel/calibration/util.py
@@ -16,6 +16,8 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
+from pyxel.inputs.loader import load_datacube, load_image
+
 #  from pyxel.pipelines import Processor
 
 __all__ = [
@@ -84,7 +86,6 @@ def read_single_data(filename: Path) -> np.ndarray:
         TBW.
     """
     # Late import to avoid circular import
-    from pyxel.inputs.loader import load_image
 
     data = load_image(filename)
 
@@ -93,6 +94,41 @@ def read_single_data(filename: Path) -> np.ndarray:
     #     raise OSError(f"Input file '{filename}' can not be read by Pyxel.")
 
     return data
+
+
+def read_single_datacube(filename: Path) -> np.ndarray:
+    """Read a numpy array from a FITS or NPY file.
+
+    Parameters
+    ----------
+    filename : Path
+
+    Returns
+    -------
+    array
+        TBW.
+    """
+
+    data = load_datacube(filename)
+
+    # # TODO: Is it the right manner ?
+    # if data is None:
+    #     raise OSError(f"Input file '{filename}' can not be read by Pyxel.")
+
+    return data
+
+
+def read_datacubes(filenames: t.Sequence[Path]) -> t.Sequence[np.ndarray]:
+    """Read numpy array(s) from several FITS or NPY files.
+
+    :param filenames:
+    :return:
+    """
+    output = [
+        read_single_datacube(Path(filename)) for filename in filenames
+    ]  # type: t.Sequence[np.ndarray]
+
+    return output
 
 
 def read_data(filenames: t.Sequence[Path]) -> t.Sequence[np.ndarray]:
@@ -111,22 +147,48 @@ def read_data(filenames: t.Sequence[Path]) -> t.Sequence[np.ndarray]:
 # TODO: Create unit tests for this function
 def list_to_slice(
     input_list: t.Optional[t.Sequence[int]] = None,
-) -> t.Union[slice, t.Tuple[slice, slice]]:
+) -> t.Union[t.Tuple[slice, slice], t.Tuple[slice, slice, slice]]:
     """TBW.
 
     :return:
     """
     if not input_list:
-        return slice(None)
-
-    if len(input_list) == 2:
-        return slice(input_list[0], input_list[1])
+        return slice(None), slice(None)
 
     elif len(input_list) == 4:
         return slice(input_list[0], input_list[1]), slice(input_list[2], input_list[3])
 
+    elif len(input_list) == 6:
+        return (
+            slice(input_list[0], input_list[1]),
+            slice(input_list[2], input_list[3]),
+            slice(input_list[4], input_list[5]),
+        )
+
     else:
-        raise ValueError("Fitting range should have 2 or 4 values")
+        raise ValueError("Fitting range should have 4 or 6 values")
+
+
+# TODO: Create unit tests for this function
+def list_to_3d_slice(
+    input_list: t.Optional[t.Sequence[int]] = None,
+) -> t.Tuple[slice, slice, slice]:
+    """TBW.
+
+    :return:
+    """
+    if not input_list:
+        return slice(None), slice(None), slice(None)
+
+    elif len(input_list) == 6:
+        return (
+            slice(input_list[0], input_list[1]),
+            slice(input_list[2], input_list[3]),
+            slice(input_list[4], input_list[5]),
+        )
+
+    else:
+        raise ValueError("Fitting range should have 6 values")
 
 
 # TODO: Write unit tests for this function
@@ -134,15 +196,16 @@ def check_ranges(
     target_fit_range: t.Sequence[int],
     out_fit_range: t.Sequence[int],
     rows: int,
-    cols: t.Optional[int] = None,
+    cols: int,
+    readout_times: t.Optional[int] = None,
 ) -> None:
     """TBW."""
     if target_fit_range:
-        if len(target_fit_range) not in (2, 4):
+        if len(target_fit_range) not in (4, 6):
             raise ValueError
 
         if out_fit_range:
-            if len(out_fit_range) not in (2, 4):
+            if len(out_fit_range) not in (4, 6):
                 raise ValueError
 
             if (target_fit_range[1] - target_fit_range[0]) != (
@@ -161,16 +224,32 @@ def check_ranges(
                         "Fitting ranges have different lengths in 2nd dimension"
                     )
 
-        for i in [0, 1]:
-            # TODO: It could be refactor in a more pythonic way
+            if len(target_fit_range) == 6 and len(out_fit_range) == 6:
+                if (target_fit_range[5] - target_fit_range[4]) != (
+                    out_fit_range[5] - out_fit_range[4]
+                ):
+                    raise ValueError(
+                        "Fitting ranges have different lengths in third dimension"
+                    )
+
+        # for i in [0, 1]:
+        #     # TODO: It could be refactor in a more pythonic way
+        #     if not (0 <= target_fit_range[i] <= rows):
+        #         raise ValueError("Value of target fit range is wrong")
+
+        for i in [-3, -4]:
             if not (0 <= target_fit_range[i] <= rows):
                 raise ValueError("Value of target fit range is wrong")
 
-        if len(target_fit_range) == 4:
-            if cols is None:
-                raise ValueError("Target data is not a 2 dimensional array")
+        for i in [-1, 2]:
+            if not (0 <= target_fit_range[i] <= cols):
+                raise ValueError("Value of target fit range is wrong")
 
-            for i in [2, 3]:
+        if len(target_fit_range) == 6:
+            if readout_times is None:
+                raise ValueError("Target data is not a 3 dimensional array")
+
+            for i in [0, 1]:
                 # TODO: It could be refactor in a more pythonic way (this is optional)
-                if not (0 <= target_fit_range[i] <= cols):
+                if not (0 <= target_fit_range[i] <= readout_times):
                     raise ValueError("Value of target fit range is wrong")